+++ lib/vhcblade_theme.dart
@@ -1,4 +1,5 @@
 library vhcblade_theme;
 
+export 'src/additional_theme.dart';
 export 'src/main_theme.dart';
 export 'src/model.dart';